Added @params to function calls, removed @param related to callback that is no longer used, moved chat related interfaces out of interfaces.ts into chat.ts
