+++ packages/teams-js/src/public/index.ts
@@ -22,7 +22,6 @@ export {
   DialogInfo,
   TeamInformation,
   FileOpenPreference,
-  OpenConversationRequest,
   LocaleInfo,
   FrameInfo,
 } from './interfaces';
