+++ packages/teams-js/src/public/interfaces.ts
@@ -578,137 +578,6 @@ export interface DialogInfo {
 export type TaskInfo = DialogInfo;
 
 /**
- *
- * @internal
- */
-interface OpenChatRequest {
-  /**
-   * @hidden
-   * The message to send when opening chat
-   */
-  message?: string;
-}
-
-/**
- * @hidden
- * Hide from docs.
- * ------
- *
- * @internal
- */
-export interface OpenSingleChatRequest extends OpenChatRequest {
-  /**
-   * @hidden
-   * User's UPN to open chat with
-   */
-  user: string;
-}
-
-/**
- * @hidden
- * Hide from docs.
- * ------
- *
- * @internal
- */
-export interface OpenGroupChatRequest extends OpenChatRequest {
-  /**
-   * @hidden
-   * Array containing UPNs of users to open chat with
-   */
-  users: string[];
-  /**
-   * @hidden
-   * The display name of a conversation for 3 or more users
-   */
-  topic?: string;
-}
-
-/**
- * @hidden
- * Hide from docs.
- * ------
- *
- * @internal
- */
-export interface OpenConversationRequest {
-  /**
-   * @hidden
-   * The Id of the subEntity where the conversation is taking place
-   */
-  subEntityId: string;
-
-  /**
-   * @hidden
-   * The title of the conversation
-   */
-  title: string;
-
-  /**
-   * @hidden
-   * The Id of the conversation. This is optional and should be specified whenever a previous conversation about a specific sub-entity has already been started before
-   */
-  conversationId?: string;
-
-  /**
-   * @hidden
-   * The Id of the channel. This is optional and should be specified whenever a conversation is started or opened in a personal app scope
-   */
-  channelId?: string;
-
-  /**
-   * @hidden
-   * The entity Id of the tab
-   */
-  entityId: string;
-
-  /**
-   * @hidden
-   * A function that is called once the conversation Id has been created
-   */
-  onStartConversation?: (conversationResponse: ConversationResponse) => void;
-
-  /**
-   * @hidden
-   * A function that is called if the pane is closed
-   */
-  onCloseConversation?: (conversationResponse: ConversationResponse) => void;
-}
-
-/**
- * @hidden
- * Hide from docs.
- * ------
- *
- * @internal
- */
-export interface ConversationResponse {
-  /**
-   * @hidden
-   * The Id of the subEntity where the conversation is taking place
-   */
-  subEntityId: string;
-
-  /**
-   * @hidden
-   * The Id of the conversation. This is optional and should be specified whenever a previous conversation about a specific sub-entity has already been started before
-   */
-  conversationId?: string;
-
-  /**
-   * @hidden
-   * The Id of the channel. This is optional and should be specified whenever a conversation is started or opened in a personal app scope
-   */
-  channelId?: string;
-
-  /**
-   * @hidden
-   * The entity Id of the tab
-   */
-  entityId?: string;
-}
-
-/**
  * @hidden
  * Hide from docs.
  *
